Importing collection to publication.js
In order for a new collection be made in mongo, there needs to be one document created in the collection. Here is perfect to place the first Name.insert({_id: 'example-testing'}); to get it started. The insert must be deleted later becasue it will keep inserting the same thing upon app reload